requirements.txt
requests
kafka_python